+++ back-ear/src/main/java/model/finance/record/FinanceRecordTableRow.java
@@ -4,7 +4,6 @@ import lombok.*;
 import model.IdNameObj;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -14,7 +13,7 @@ import java.util.Date;
 public class FinanceRecordTableRow implements Serializable {
     public String id;
     public IdNameObj account;
-    public BigDecimal amount;
+    public Double amount;
     public IdNameObj category;
     public String description;
     public Date date;
